Add install cell and update imports in notebook
