+++ dp_pipeline/models.py
@@ -1,0 +1,73 @@
+from __future__ import annotations
+
+try:
+    import torch
+    import torch.nn as nn
+    import torch.nn.functional as F
+except ImportError:  # pragma: no cover - handled at runtime
+    torch = None  # type: ignore
+    nn = None  # type: ignore
+    F = None  # type: ignore
+
+
+def resnet20() -> nn.Module:
+    """Returns a small ResNet-20 model for CIFAR-10."""
+    if torch is None or nn is None or F is None:
+        raise RuntimeError("PyTorch must be installed to construct models.")
+
+    class BasicBlock(nn.Module):
+        expansion = 1
+
+        def __init__(self, in_planes, planes, stride=1):
+            super().__init__()
+            self.conv1 = nn.Conv2d(in_planes, planes, kernel_size=3, stride=stride, padding=1, bias=False)
+            self.bn1 = nn.BatchNorm2d(planes)
+            self.conv2 = nn.Conv2d(planes, planes, kernel_size=3, stride=1, padding=1, bias=False)
+            self.bn2 = nn.BatchNorm2d(planes)
+
+            self.shortcut = nn.Sequential()
+            if stride != 1 or in_planes != planes:
+                self.shortcut = nn.Sequential(
+                    nn.Conv2d(in_planes, planes, kernel_size=1, stride=stride, bias=False),
+                    nn.BatchNorm2d(planes),
+                )
+
+        def forward(self, x):
+            out = F.relu(self.bn1(self.conv1(x)))
+            out = self.bn2(self.conv2(out))
+            out += self.shortcut(x)
+            out = F.relu(out)
+            return out
+
+    class ResNet(nn.Module):
+        def __init__(self, block, num_blocks, num_classes=10):
+            super().__init__()
+            self.in_planes = 16
+
+            self.conv1 = nn.Conv2d(3, 16, kernel_size=3, stride=1, padding=1, bias=False)
+            self.bn1 = nn.BatchNorm2d(16)
+            self.layer1 = self._make_layer(block, 16, num_blocks[0], stride=1)
+            self.layer2 = self._make_layer(block, 32, num_blocks[1], stride=2)
+            self.layer3 = self._make_layer(block, 64, num_blocks[2], stride=2)
+            self.avgpool = nn.AdaptiveAvgPool2d((1, 1))
+            self.linear = nn.Linear(64, num_classes)
+
+        def _make_layer(self, block, planes, num_blocks, stride):
+            strides = [stride] + [1] * (num_blocks - 1)
+            layers = []
+            for s in strides:
+                layers.append(block(self.in_planes, planes, s))
+                self.in_planes = planes * block.expansion
+            return nn.Sequential(*layers)
+
+        def forward(self, x):
+            out = F.relu(self.bn1(self.conv1(x)))
+            out = self.layer1(out)
+            out = self.layer2(out)
+            out = self.layer3(out)
+            out = self.avgpool(out)
+            out = torch.flatten(out, 1)
+            out = self.linear(out)
+            return out
+
+    return ResNet(BasicBlock, [3, 3, 3])
